requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.30
python-dotenv>=1.0
google-genai>=0.3
pydantic>=2.7
python-multipart>=0.0.9
